feat: P2 v0.2 — neatmesh integration + performance benchmarking
- Integrate neatmesh supplementary metrics into NativeMeshChecker.run()
  - Add _run_neatmesh_from_polyMesh() method for polyMesh→VTK→neatmesh pipeline
  - Graceful fallback when pyvista can't read polyMesh format (non-critical path)
  - Log supplementary metrics: non-orthogonality, adjacency ratios, cell type distribution

- Implement scripts/benchmark_test_cases.py for performance analysis
  - Iterate through 26 test cases (9 basic + 8 advanced + legacy cases)
  - Run each with --quality draft for baseline performance
  - Generate JSON + Markdown reports with execution time, cell count, mesh OK verdict
  - Statistics: min/max/avg/median timing and cell counts

- Create P2_V0.2_IMPLEMENTATION.md documenting changes and next steps

Test Results: 1016 passed, 12 skipped, 2 warnings ✅
Benchmark Test: sphere_watertight.stl executes in 3.71s ✅

+++ core/analyzer/file_reader.py
@@ -195,7 +195,8 @@ def _load_via_gmsh(path: Path, fmt: str) -> trimesh.Trimesh:
         if not tri_faces:
             raise ValueError(f"gmsh 메쉬에서 삼각형 요소를 찾지 못했습니다: {path}")
 
-        faces = np.vstack(tri_faces).astype(np.int32)
+        # 인덱스는 플랫폼 네이티브 정수 폭으로 유지해 대형 태그 손실을 방지한다.
+        faces = np.vstack(tri_faces).astype(np.intp, copy=False)
         mesh = trimesh.Trimesh(vertices=vertices, faces=faces, process=True)
 
         if len(mesh.faces) == 0:
